+++ src/services/filters-parser.js
@@ -1,8 +1,9 @@
-import { BaseFiltersParser, BaseOperatorDateParser } from 'forest-express';
+import { BaseFiltersParser, BaseOperatorDateParser, Schemas } from 'forest-express';
+import { getReferenceField } from '../utils/query';
 import Operators from '../utils/operators';
 import { NoMatchingOperatorError } from './errors';
 
-function FiltersParser(timezone, options) {
+function FiltersParser(modelSchema, timezone, options) {
   this.OPERATORS = new Operators(options);
   this.operatorDateParser = new BaseOperatorDateParser({ operators: this.OPERATORS, timezone });
 
@@ -121,7 +122,13 @@ function FiltersParser(timezone, options) {
     }
   };
 
-  this.formatField = field => (field.includes(':') ? `$${field.replace(':', '.')}$` : field);
+  this.formatField = (field) => {
+    if (field.includes(':')) {
+      const [associationName, fieldName] = field.split(':');
+      return `$${getReferenceField(Schemas.schemas, modelSchema, associationName, fieldName)}$`;
+    }
+    return field;
+  };
 
   // NOTICE: Look for a previous interval condition matching the following:
   //         - If the filter is a simple condition at the root the check is done right away.