+++ src/services/query-builder.js
@@ -1,5 +1,6 @@
 import _ from 'lodash';
 import { Schemas } from 'forest-express';
+import { getReferenceField } from '../utils/query';
 import Orm from '../utils/orm';
 import Database from '../utils/database';
 
@@ -64,12 +65,13 @@ function QueryBuilder(model, opts, params) {
       if (params.sort.indexOf('.') !== -1) {
         // NOTICE: Sort on the belongsTo displayed field
         const [associationName, fieldName] = params.sort.split('.');
-        const schemaField = (aliasSchema || schema).fields
-          .find(field => field.field === associationName);
-        const [tableName] = schemaField.reference.split('.');
-        const associationSchema = Schemas.schemas[tableName];
-        const belongsToColumnName = Orm.getColumnName(associationSchema, fieldName);
-        return [[opts.sequelize.col(`${associationName}.${belongsToColumnName}`), order]];
+        const column = getReferenceField(
+          Schemas.schemas,
+          (aliasSchema || schema),
+          associationName,
+          fieldName,
+        );
+        return [[opts.sequelize.col(column), order]];
       } else if (aliasName) {
         return [[opts.sequelize.col(`${aliasName}.${Orm.getColumnName(aliasSchema, params.sort)}`), order]];
       }